+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.140.0] - Not released
+### Added
+- Quick Links (TOC) on the Appearance tab of the Settings page.
 
 ## [1.139.0] - 2025-08-02
 ### Changed